### 2025-02-24 65b554b
Fix YAML header in paper.md
