### 2025-03-17 #7810 @Kevin-Mattheus-Moerman
- [ ] Please study the above reference check. Please check if you are citing all these or if the .bib file should be purged to only include the works you do cite.
