### 2025-03-17 925b921
Purge redundant refs in paper.bib
