+++ JOSS/paper.bib
@@ -33,23 +33,7 @@
 	langid = {english},
 }
 
-@article{sonti_role_1989,
-	title = {Role of gene duplications in the adaptation of Salmonella typhimurium to growth on limiting carbon sources.},
-	volume = {123},
-	rights = {https://academic.oup.com/journals/pages/open\_access/funder\_policies/chorus/standard\_publication\_model},
-	issn = {1943-2631},
-	url = {https://academic.oup.com/genetics/article/123/1/19/5998750},
-	doi = {10.1093/genetics/123.1.19},
-	abstract = {Abstract 
-            Duplication-containing cells are selected when growth of Salmonella typhimurium is limited by the availability of any one of several carbon and energy sources. Under conditions of extreme starvation, growth occurs almost exclusively in the duplication-containing fraction of the population. Cells with duplications of one large segment of the chromosome are repeatedly selected regardless of which of these carbon sources limits growth. The duplicated chromosomal segment encodes the transport systems for all of these carbon sources. This duplication is not selected during growth on a carbon source for which the permease is not included within the duplication segment. This suggests that the growth advantage conferred by the duplication may be due to increased transport of the limiting carbon source. Inclusion of the permease alone is not sufficient to explain the growth advantage of the duplications, since other common duplications that include the permease are not selected.},
-	pages = {19--28},
-	number = {1},
-	journaltitle = {Genetics},
-	author = {Sonti, R V and Roth, J R},
-	urldate = {2025-01-31},
-	date = {1989-09-01},
-	langid = {english},
-}
+
 
 @article{seemann_prokka_2014,
 	title = {Prokka: rapid prokaryotic genome annotation},
@@ -155,39 +139,6 @@
 }
 
 
-@article{xie_isescan_2017,
-	title = {{ISEScan}: automated identification of insertion sequence elements in prokaryotic genomes},
-	volume = {33},
-	rights = {https://academic.oup.com/journals/pages/about\_us/legal/notices},
-	issn = {1367-4803, 1367-4811},
-	url = {https://academic.oup.com/bioinformatics/article/33/21/3340/3930124},
-	doi = {10.1093/bioinformatics/btx433},
-	shorttitle = {{ISEScan}},
-	abstract = {Abstract 
-             
-              Motivation 
-              The insertion sequence ({IS}) elements are the smallest but most abundant autonomous transposable elements in prokaryotic genomes, which play a key role in prokaryotic genome organization and evolution. With the fast growing genomic data, it is becoming increasingly critical for biology researchers to be able to accurately and automatically annotate {ISs} in prokaryotic genome sequences. The available automatic {IS} annotation systems are either providing only incomplete {IS} annotation or relying on the availability of existing genome annotations. Here, we present a new {IS} elements annotation pipeline to address these issues. 
-             
-             
-              Results 
-              {ISEScan} is a highly sensitive software pipeline based on profile hidden Markov models constructed from manually curated {IS} elements. {ISEScan} performs better than existing {IS} annotation systems when tested on prokaryotic genomes with curated annotations of {IS} elements. Applying it to 2784 prokaryotic genomes, we report the global distribution of {IS} families across taxonomic clades in Archaea and Bacteria. 
-             
-             
-              Availability and implementation 
-              {ISEScan} is implemented in Python and released as an open source software at https://github.com/xiezhq/{ISEScan}. 
-             
-             
-              Supplementary information 
-              Supplementary data are available at Bioinformatics online.},
-	pages = {3340--3347},
-	number = {21},
-	journaltitle = {Bioinformatics},
-	author = {Xie, Zhiqun and Tang, Haixu},
[…]
