### 2025-03-17 #7810 @ira-zibbu
Hi @Kevin-Mattheus-Moerman ,

- I have added a `contributing.md` file, and linked to it in the README
- I have purged redundant references in `paper.bib`

Thank you!
