### 2025-04-12 4e00638
Remove unused refs from paper.bib
