+++ JOSS/paper.bib
@@ -7,21 +7,21 @@
 	url = {https://academic.oup.com/bioinformatics/article/33/21/3340/3930124},
 	doi = {10.1093/bioinformatics/btx433},
 	shorttitle = {{ISEScan}},
-	abstract = {Abstract 
-             
-              Motivation 
-              The insertion sequence ({IS}) elements are the smallest but most abundant autonomous transposable elements in prokaryotic genomes, which play a key role in prokaryotic genome organization and evolution. With the fast growing genomic data, it is becoming increasingly critical for biology researchers to be able to accurately and automatically annotate {ISs} in prokaryotic genome sequences. The available automatic {IS} annotation systems are either providing only incomplete {IS} annotation or relying on the availability of existing genome annotations. Here, we present a new {IS} elements annotation pipeline to address these issues. 
-             
-             
-              Results 
-              {ISEScan} is a highly sensitive software pipeline based on profile hidden Markov models constructed from manually curated {IS} elements. {ISEScan} performs better than existing {IS} annotation systems when tested on prokaryotic genomes with curated annotations of {IS} elements. Applying it to 2784 prokaryotic genomes, we report the global distribution of {IS} families across taxonomic clades in Archaea and Bacteria. 
-             
-             
-              Availability and implementation 
-              {ISEScan} is implemented in Python and released as an open source software at https://github.com/xiezhq/{ISEScan}. 
-             
-             
-              Supplementary information 
+	abstract = {Abstract
+
+              Motivation
+              The insertion sequence ({IS}) elements are the smallest but most abundant autonomous transposable elements in prokaryotic genomes, which play a key role in prokaryotic genome organization and evolution. With the fast growing genomic data, it is becoming increasingly critical for biology researchers to be able to accurately and automatically annotate {ISs} in prokaryotic genome sequences. The available automatic {IS} annotation systems are either providing only incomplete {IS} annotation or relying on the availability of existing genome annotations. Here, we present a new {IS} elements annotation pipeline to address these issues.
+
+
+              Results
+              {ISEScan} is a highly sensitive software pipeline based on profile hidden Markov models constructed from manually curated {IS} elements. {ISEScan} performs better than existing {IS} annotation systems when tested on prokaryotic genomes with curated annotations of {IS} elements. Applying it to 2784 prokaryotic genomes, we report the global distribution of {IS} families across taxonomic clades in Archaea and Bacteria.
+
+
+              Availability and implementation
+              {ISEScan} is implemented in Python and released as an open source software at https://github.com/xiezhq/{ISEScan}.
+
+
+              Supplementary information
               Supplementary data are available at Bioinformatics online.},
 	pages = {3340--3347},
 	number = {21},
@@ -42,9 +42,9 @@
 	url = {https://academic.oup.com/bioinformatics/article/30/14/2068/2390517},
 	doi = {10.1093/bioinformatics/btu153},
 	shorttitle = {Prokka},
-	abstract = {Abstract 
[…]
