### 2025-04-12 b9e582f
Add missing comma in paper.bib
