+++ JOSS/paper.bib
@@ -130,7 +130,7 @@
 	langid = {english},
 }
 
-@article{goel_syri_2019
+@article{goel_syri_2019,
 	title = {{SyRI}: finding genomic rearrangements and local sequence differences from whole-genome assemblies},
 	volume = {20},
 	issn = {1474-760X},
