### 2025-04-14 #7810 @abhishektiwari
👋 @dr-joe-wirth @JeanMainguy @MmasterT, would any of you be willing to review this submission for JOSS?  The submission under consideration is "seabreeze: A Pipeline for Analyzing Structural Variation Between Bacterial Genome Assemblies".

Based on your experience, we think you might be able to provide a great review of this submission. If not, could you recommend any potential reviewers?

Thanks
Abhi
