### 2025-04-16 #8065 @editorialbot
[REVIEW]: seabreeze: A Pipeline for Analyzing Structural Variation Between Bacterial Genome Assemblies
**Submitting author:** @ira-zibbu (<a href="http://orcid.org/0009-0003-3930-5397">Ira Zibbu</a>)
**Repository:** https://github.com/barricklab/seabreeze
**Branch with paper.md** (empty if default branch):
**Version:** v.1.1.0
