### 2025-04-24 #8065 @dr-joe-wirth
Three authors are listed on the manuscript, but @ira-zibbu is [the only contributor to the repository](https://github.com/barricklab/seabreeze/graphs/contributors). The JOSS reviewer instructions state:

> As discussed in the [submission guidelines for authors](https://joss.readthedocs.io/en/latest/submitting.html#authorship), authorship is a complex topic with different practices in different communities. Ultimately, the authors themselves are responsible for deciding which contributions are sufficient for co-authorship, although JOSS policy is that purely financial contributions are not considered sufficient. Your job as a reviewer is to check that the list of authors appears reasonable, and if it’s not obviously complete/correct, to raise this as a question during the review.

It is not immediately obvious to me why Claus O. Wilke and Jeffrey Barrick are listed as authors on the manuscript. Can you please explain the rationale for including Claus O. Wilke and Jeffrey Barrick as authors?
