### 2025-04-24 barricklab/seabreeze#4 @dr-joe-wirth
Add instructions for running unit tests
These instructions should be included somewhere on the repo. I suggest including them in the contributing doc as well.

This is part of the review at https://github.com/openjournals/joss-reviews/issues/8065
