### 2025-04-24 32a0226
Fix typo in paper.md
