### 2025-04-24 #8065 @ira-zibbu
Hi @dr-joe-wirth,

Thank you for your comments. I have corrected the typo in the paper as requested.

Jeffrey Barrick contributed to the conceptualization, funding acquisition, development of methodology, supervision, validation of the software, and writing (review & editing). Claus Wilke contributed to the development of the methodology, supervision, and writing (review & editing).  I had extensive conversations with both Jeffrey Barrick and Claus Wilke about the high-level design of the software, specific implementation details, and addressing various issues that arose during development.
