### 2025-04-25 barricklab/seabreeze#5 @JeanMainguy
Adding unit tests to python CI
Hi,
I noticed that the current Python CI workflow is only building a conda environment from the `environment.yml` file.
Would it be possible to extend it to also run the unit tests? That would help verify the code behaves as expected.

This is part of the review at: https://github.com/openjournals/joss-reviews/issues/8065
