### 2025-04-25 barricklab/seabreeze#5 @JeanMainguy
The CI could also run the seabreeze commands shown in the tutorial section, and potentially check the generated output files. That would help ensure everything runs smoothly from end to end.
