### 2025-04-25 barricklab/seabreeze#6 @JeanMainguy
Running all seabreeze commands at once
Hi,
I was able to complete the tutorial. it's well done, detailed, and walks through everything clearly.

I was wondering: is there a way to run a single seabreeze command that produces all the outputs at once, rather than running multiple steps separately?

This is part of the review at: https://github.com/openjournals/joss-reviews/issues/8065
