### 2025-04-25 barricklab/seabreeze#7 @JeanMainguy
Improve integration of `plostr` to avoid code duplication
Hi,

I noticed that the source code of the `plostr` tool has been directly copied into this repository with some modifications.

Since `plostr` is licensed under MIT, this is perfectly legal. However, including and modifying the full codebase directly can make future maintenance more difficult as it prevents seabreeze from easily benefiting from updates or bug fixes from `plostr`.

A more maintainable approach could be:
- Contributing the modifications back to the original `plostr` repository, if they are general enough to be useful to others, or
- Using `plostr` as a Python package and applying the changes in a dedicated wrapper script or function. `plostr` is written in Python, so I guess it should be feasible to integrate it in a more modular way without duplicating the entire codebase.

Is there a particular reason why the full `plostr` code was included in the repository ?

This is part of the review at: https://github.com/openjournals/joss-reviews/issues/8065
