### 2025-05-02 barricklab/seabreeze#4 @ira-zibbu
That would be useful. I will add instructions in the documentation and in `contributing.md`.
