### 2025-05-02 barricklab/seabreeze#5 @ira-zibbu
Hi @JeanMainguy ,

Thanks for raising the issue. Seems like when I updated the CI workflow, I did not add back the step to run `pytest` to execute the unit tests. Will update the CI to reflect this.

I agree, adding additional tests in CI that run the seabreeze commands would be helpful. Will work on implementing these tests.
