### 2025-05-02 barricklab/seabreeze#6 @ira-zibbu
Hi @JeanMainguy ,

Currently, seabreeze does not support a command to generate all of the output at once, but implementing a master command is straightforward. I agree that it would be useful.
