### 2025-05-02 barricklab/seabreeze#7 @ira-zibbu
Hi @JeanMainguy,

The original version of `plotsr` did not support depicting deletion mutations in the synteny plots, hence I started out by forking their repository and making modifications to add this new functionality. My fork of `plotsr` existed as a submodule inside the seabreeze repository. I felt that cloning the seabreeze repository, and then having to clone the fork of plotsr as a submodule inside it, was complicated as an installation step for the user. I duplicated my fork of plotsr inside seabreeze so that all of it could be released together as a single GitHub release.

I agree with you that this solution is not optimal, as this does not allow for updates or fixes in plotsr to be incorporated. I will submit a pull request to the original plotsr repository with the new feature. If my PR is accepted and the new merged version is available on bioconda, I will update seabreeze to use conda to install plotsr instead.
