### 2025-05-05 b54ebb2
Add pyproject.toml, set up main CLI with click
