+++ src/seabreeze/syri2gd.py
@@ -1,0 +1,140 @@
+#!/usr/bin/env python3
+#Author: ira-zibbu
+#run from 06_Syri_Output
+#This script converts all the deletions, inversions and duplications in a syri.out file to lines in genome diff format used by breseq
+
+''' imports '''
+
+import os
+import pandas as pd
+import numpy as np
+import argparse
+
+''' fetch arguments '''
+
+parser = argparse.ArgumentParser(description='syri2gd.py, a script to convert a syri.out file to genome diff format used by breseq')
+parser.add_argument('--syri', help='syri.out file')
+parser.add_argument('--output', help='name of output gd file')
+parser.add_argument('--inversion', action='store_true', help='include inversions in gd file')
+parser.add_argument('--deletion', action='store_true', help='include deletions in gd file')
+parser.add_argument('--amplification', action='store_true', help='include amplifications in gd file')
+parser.add_argument('--verbose', '-v', action='store_true', help='Enable verbose mode')
+
+
+def syri2gd(syri,output,inversion,deletion,amplification):
+    
+    df_syri = pd.read_table(syri,delimiter='\t', header=None) 
+    chr_id=df_syri.iloc[0,0] #reference chromosome ID/ contig ID
+    df_syri_del = df_syri[df_syri.iloc[:,10].str.contains('DEL')]
+    # df_syri_inv = df_syri[df_syri.iloc[:,10.str.contains('INV')]
+    # df_syri_amp = df_syri[df_syri.iloc[:,10].str.contains('DUP|INVDP')]
+    df_syri_inv = df_syri[df_syri.iloc[:, 10] == 'INV']
+    df_syri_amp = df_syri[df_syri.iloc[:, 10].isin(['DUP', 'INVDP','CPG'])]
+
+
+    df_syri_amp = df_syri_amp.reset_index(drop=True)
+
+    df_gd_del=pd.DataFrame(np.nan, index=range(len(df_syri_del)), columns=range(6))
+    df_gd_inv=pd.DataFrame(np.nan, index=range(len(df_syri_inv)), columns=range(6))
+    df_gd_amp=pd.DataFrame(np.nan, index=range(len(df_syri_amp)), columns=range(7))
+
+
+    header_line="#=GENOME_DIFF\t1.0"
+    with open(output, 'w') as file:
+        file.write(header_line + '\n')  # Write the comment header line
+    row_idx = 0
+
+    if deletion: # add genome diff lines for deletions
+
+        for row_idx in range(len(df_syri_del)):
+            
+            start_coords=df_syri_del.iloc[row_idx,1]
+            end_coords=df_syri_del.iloc[row_idx,2]
+            del_length=int(end_coords)-int(start_coords)
+
+            if del_length <= 0: # skip any entry which has 0 length
+                continue
+
+            df_gd_del.iloc[row_idx,0]="DEL"
+            df_gd_del.iloc[row_idx,1]=str(row_idx+1)
+            df_gd_del.iloc[row_idx,2]="."
+            df_gd_del.iloc[row_idx,3]=chr_id
+           
+            df_gd_del.iloc[row_idx,4]=str(start_coords)
+            df_gd_del.iloc[row_idx,5]=str(del_length)
+        
+        df_gd_del.dropna(inplace=True)
+        df_gd_del.reset_index(drop=True, inplace=True)
+
+        with open(output, 'a') as file:
+            df_gd_del.to_csv(file, sep='\t', index=False, header=False)
+    
+    if inversion: # add genome diff lines for inversions
+
+        index_list=range(row_idx+2,(len(df_syri_inv)+row_idx+2),1) # to continue the indices from the gd file
+        for row_idx in range(len(df_syri_inv)):
+            
[…]
