### 2025-05-05 2a71cf5
Add seabreeze cli script to call snakemake subcommands, use non-seabreeze dir as working dir
