### 2025-05-08 cb191b0
Create separate commands for batch and quick run. Add rules for structural variant prediction. Add test for structural variant prediction
