### 2025-05-08 078992c
Complete migration of snakefile, add option to run just two genomes
