+++ src/seabreeze/seabreeze.py
@@ -13,6 +13,8 @@ import os
 import logging
 import coloredlogs
 from seabreeze import __version__
+import shutil
+import csv
 
 # set up logger
 coloredlogs.install(level='DEBUG')
@@ -56,7 +58,10 @@ def cli():
         [
             "analyse_genome_sizes",
             "predict_IS_elements",
-            "predict_structural_variants"
+            "predict_structural_variants",
+            "predict_replichore_balance",
+            "predict_SV_mechanism",
+            "annotate_SV_regions"
         ]
     ),
 )
@@ -83,18 +88,6 @@ def cli():
     default="./ori_dif_coords.csv",
 )
 
-# @click.option(
-#     "--ancestor",
-#     type=click.Path(dir_okay=True, writable=True, resolve_path=True),
-#     help="Path to FASTA file of the ancestor or reference sequence",
-# )
-
-# @click.option(
-#     "--assembly",
-#     type=click.Path(dir_okay=True, writable=True, resolve_path=True),
-#     help="Path to FASTA file of the assembly or query sequence",
-# )
-
 @click.option(
     "--threads",
     type=int,
@@ -107,9 +100,33 @@ def cli():
 
 @click.argument("snakemake_args", nargs=-1, type=click.UNPROCESSED)
 
-def run_smk_workflow(workflow,dir,data,oridif,masked,threads,snakemake_args):
+def run_smk_batch_workflow(workflow,dir,data,oridif,masked,threads,snakemake_args):
 
     """ Determine snakemake command to run based on supplied args """
+
+    logger.info(f"==================================")
+    logger.info(f"seabrezee version: {__version__}")
+    logger.info(f"==================================")
+
+    # make the working directory if it does not already exist
+    if not os.path.exists(dir):
+        logger.info(f"{dir} does not exist. Creating the directory..")
+        try:
+            os.makedirs(dir)
+        except OSError as e:
+            logger.critical(f"Unable to create {dir}")
+            logger.critical(e)
+
+    # verify that data.csv exists
+    if not os.path.exists(data):
+        logger.critical(f"{data} does not exist. Aborting")
+        exit(1)
+
+    # ori and dif sequences are required to predict replichore balance:
+    if workflow == "predict_replichore_balance":
+        if not os.path.exists(oridif):
+            logger.critical(f"{oridif} does not exist. Aborting")
+            exit(1)
 
     os.chdir(dir)
     logger.info(f"Switch working directory to {os.getcwd()}")
@@ -124,7 +141,201 @@ def run_smk_workflow(workflow,dir,data,oridif,masked,threads,snakemake_args):
 
     seabreeze_package_path = importlib.resources.files(seabreeze)
 
-    config_options=f"data={data} threads={threads}"
+    config_options=f"data={data} threads={threads} oridif={oridif}"
+    cmd=()
+
+    cmd = (
+        "snakemake --snakefile {snakefile} --use-conda --cores {cores} {workflow} --config {config}"
+        " {args} "
+    ).format(
+        snakefile=get_snakefile(),
+        cores=threads,
+        workflow=workflow,
+        args=" ".join(snakemake_args),
+        config=config_options
+    )
+    logger.debug("Executing: %s" % cmd)
+
+    try:
+        subprocess.check_call(cmd, shell=True)
+    except subprocess.CalledProcessError as e:
+        logger.critical(e)
+        exit(1)
+
+# ---- SINGLE RUN MODE ----
+
+@cli.command(
+    "run",
+    context_settings=dict(ignore_unknown_options=True),
+    short_help="Run seabreeze for a single sample pair",
+)
+
+@click.argument(
+    "workflow",
+    type=click.Choice(
+        [
+            "analyse_genome_sizes",
[…]
