### 2025-05-12 barricklab/seabreeze#7 @ira-zibbu
Hi,

I have submitted a[ PR to the original plotsr repository](https://github.com/schneebergerlab/plotsr/pull/100). Waiting to hear back from the authors to see if they accept the PR
