### 2025-05-14 984edbf
Add workflow tests for seabreeze run and batch. Update ci.yaml
