+++ test/seabreeze_batch_workflow_test.py
@@ -1,0 +1,140 @@
+"""
+Master test to see if the seabreeze workflow pipelines run
+"""
+
+import subprocess
+import pytest
+import csv
+import os
+import shutil
+
+def read_csv_as_dicts(path):
+    with open(path, newline='') as f:
+        return list(csv.DictReader(f))
+
+def test_batch_analyse_genome_sizes():
+
+    command = ["seabreeze", "batch", "analyse_genome_sizes", "--data", "test/seabreeze_batch_test/data.csv", "--dir", "test/seabreeze_batch_test/"]
+    result = subprocess.run(command)
+    assert result.returncode == 0, f"Command failed with error: {result.stderr}"
+
+    path_to_template_output="test/seabreeze_batch_test/expected/genome_size_stats.csv"
+    path_to_generated_output="test/seabreeze_batch_test/04_rename_genome/genome_size_stats.csv"
+
+    # actual = read_csv_as_dicts(path_to_generated_output)
+    # expected = read_csv_as_dicts(path_to_template_output)
+
+    # assert actual == expected
+
+def test_batch_predict_IS_elements():
+
+    command = ["seabreeze", "batch", "--dir","test/seabreeze_batch_test/","--data", "test/seabreeze_batch_test/data.csv", "predict_IS_elements"]
+
+    result = subprocess.run(command)
+    assert result.returncode == 0, f"Command failed with error: {result.stderr}"
+
+    path_to_template_output="test/seabreeze_batch_test/expected/Anc-_0gen_REL606_truncated.csv"
+    path_to_generated_output="test/seabreeze_batch_test/05_isescan_tables/Anc-_0gen_REL606_truncated.csv"
+
+    # actual = read_csv_as_dicts(path_to_generated_output)
+    # expected = read_csv_as_dicts(path_to_template_output)
+
+    # assert actual == expected
+
+def test_batch_predict_structural_variants_unmasked():
+
+    command = ["seabreeze", "batch", "--dir","test/seabreeze_batch_test/","--data", "test/seabreeze_batch_test/data.csv", "predict_structural_variants"]
+
+    result = subprocess.run(command)
+    assert result.returncode == 0, f"Command failed with error: {result.stderr}"
+
+    # path_to_template_output="test/seabreeze_batch_test/expected/Ara+1_10000gen_4530A_truncated_clean_unmasked.syri.out"
+    # path_to_generated_output="test/seabreeze_batch_test/07_syri_output/unmasked/Ara+1_10000gen_4530A_truncated/Ara+1_10000gen_4530A_truncated_clean.syri.out"
+
+    # actual = read_csv_as_dicts(path_to_generated_output)
+    # expected = read_csv_as_dicts(path_to_template_output)
+
+    # assert actual == expected
+
+def test_batch_predict_structural_variants_masked():
+
+    command = ["seabreeze", "batch", "--dir","test/seabreeze_batch_test/","--data", "test/seabreeze_batch_test/data.csv", "--masked", "predict_structural_variants"]
+
+    result = subprocess.run(command)
+    assert result.returncode == 0, f"Command failed with error: {result.stderr}"
+
+    # path_to_template_output="test/seabreeze_batch_test/expected/Ara+1_10000gen_4530A_truncated_clean_masked.syri.out"
+    # path_to_generated_output="test/seabreeze_batch_test/07_syri_output/unmasked/Ara+1_10000gen_4530A_truncated/Ara+1_10000gen_4530A_truncated_clean.syri.out"
+
+    # actual = read_csv_as_dicts(path_to_generated_output)
+    # expected = read_csv_as_dicts(path_to_template_output)
+
+    # assert actual == expected
+
[…]
