### 2025-05-14 2ac2717
Add click to environment.yml and remove draft pdf CI workflow
