### 2025-05-15 79546dd
Update seabreeze CLI with run_all command
