+++ src/seabreeze/seabreeze.py
@@ -1,7 +1,7 @@
 """
 seabreeze.py, a wrapper script around a snakemake pipeline to analyse structural variants in bacterial genomes
 Author: Ira Zibbu
-Version: 0.0.1
+Version: 0.0.3
 """
 
 """ imports """
@@ -15,6 +15,7 @@ import coloredlogs
 from seabreeze import __version__
 import shutil
 import csv
+import pandas as pd
 
 # set up logger
 coloredlogs.install(level='DEBUG')
@@ -37,6 +38,7 @@ def get_snakefile():
 @click.version_option(__version__)
 
 def cli():
+
     """
     seabreeze is a tool for comprehensively analyzing genetic variation among bacterial genomes caused by structural mutations.
     To get help, run `seabreeze --help`
@@ -61,7 +63,8 @@ def cli():
             "predict_structural_variants",
             "predict_replichore_balance",
             "predict_SV_mechanism",
-            "annotate_SV_regions"
+            "annotate_SV_regions",
+            "run_all"
         ]
     ),
 )
@@ -104,13 +107,13 @@ def run_smk_batch_workflow(workflow,dir,data,oridif,masked,threads,snakemake_arg
 
     """ Determine snakemake command to run based on supplied args """
 
-    logger.info(f"==================================")
-    logger.info(f"seabrezee version: {__version__}")
-    logger.info(f"==================================")
+    logger.debug(f"==================================")
+    logger.debug(f"seabrezee version: {__version__}")
+    logger.debug(f"==================================")
 
     # make the working directory if it does not already exist
     if not os.path.exists(dir):
-        logger.info(f"{dir} does not exist. Creating the directory..")
+        logger.debug(f"{dir} does not exist. Creating the directory..")
         try:
             os.makedirs(dir)
         except OSError as e:
@@ -122,6 +125,30 @@ def run_smk_batch_workflow(workflow,dir,data,oridif,masked,threads,snakemake_arg
         logger.critical(f"{data} does not exist. Aborting")
         exit(1)
 
+
+    df = pd.read_csv(data)
+
+    # Check if the csv has exactly two columns
+    if df.shape[1] != 2:
+        logger.critical("data.csv must have exactly two columns.")
+
+    # Check if the columns are named correctly
+    expected_columns = ['assembly', 'ancestor']
+    if list(df.columns) != expected_columns:
+        logger.critical(f"The data.csv columns must be named {expected_columns}. Found: {list(df.columns)}")
+
+    # Check if there are any duplicate entries for assemblies
+    if df['assembly'].duplicated().any():
+        logger.critical("There are duplicate entries in the assembly column.")
+
+    list_of_ancestors = df['ancestor'].tolist()
+    list_of_assemblies = df['assembly'].tolist()
+
+    for ancestor in list_of_ancestors:
+        if not(ancestor in list_of_assemblies):
+            logger.critical(f"Ancestor genome {ancestor} needs a self-to-self comparison. Please append this line to data.csv: {ancestor},{ancestor}")
+
+
     # ori and dif sequences are required to predict replichore balance:
     if workflow == "predict_replichore_balance":
         if not os.path.exists(oridif):
@@ -129,10 +156,10 @@ def run_smk_batch_workflow(workflow,dir,data,oridif,masked,threads,snakemake_arg
             exit(1)
 
     os.chdir(dir)
-    logger.info(f"Switch working directory to {os.getcwd()}")
-    logger.info(f"seabrezee version: {__version__}")
+    logger.debug(f"Switch working directory to {os.getcwd()}")
+    logger.debug(f"seabrezee version: {__version__}")
 
[…]
