+++ src/seabreeze/workflow/envs/isescan.yml
@@ -3,4 +3,4 @@ channels:
   - bioconda
   - defaults
 dependencies:
-  - isescan=1.7.2.3
+  - isescan
