### 2025-05-15 8a44de1
Delete old bin folder and update dir organization for tests
