### 2025-05-15 76c92a5
Fix issue where threads param was passed twice to snakemake, fix typo in test file
