### 2025-05-15 ade5731
Fix incorrect command call in test/seabreeze_batch_all_test.py
