### 2025-05-15 788d3c4
Add some print statements to debug missing snakefile
