### 2025-05-15 61c13cd
Update pyproject.toml to explicity include snakefile while building package
