### 2025-05-15 4a0201b
seabreeze debugging unable to find plotsr-bin
