### 2025-05-15 4cd9568
Update pyproject.toml to include src/seabreeze/plotsr/plotsr-bin while building
