+++ pyproject.toml
@@ -14,6 +14,7 @@ include-package-data = true
 
 [tool.setuptools.package-data]
 "seabreeze" = ["Snakefile", "*.smk", "*.yml"]
+"seabreeze.plotsr" = ["plotsr-bin"]
 
 [project.scripts]
 seabreeze = "seabreeze.seabreeze:main"
