### 2025-05-15 ad9dfb6
Add conda front-end options for snakemake test runs
