### 2025-05-16 b305928
Add workflow yaml files to pyproject.toml
