+++ pyproject.toml
@@ -15,6 +15,10 @@ include-package-data = true
 [tool.setuptools.package-data]
 "seabreeze" = ["Snakefile", "*.smk", "*.yml"]
 "seabreeze.plotsr" = ["plotsr-bin"]
+[tool.setuptools.package-data]
+"seabreeze" = ["Snakefile", "*.smk", "*.yml"]
+"seabreeze.plotsr" = ["plotsr-bin"]
+"seabreeze.workflow.envs" = ["*.yml", "*.yaml"]
 
 [project.scripts]
 seabreeze = "seabreeze.seabreeze:main"
