### 2025-05-22 9720872
Switch from snakemake to snakemake-minimal to avoid package conflicts
