+++ environment.yml
@@ -4,7 +4,7 @@ channels:
   - bioconda
   - defaults
 dependencies:
-  - snakemake=7.32.4
+  - snakemake-minimal=7.32.4
   - numpy=2.2.5
   - pandas=2.2.3
   - pytest=8.3.5
