### 2025-05-22 aef0d3a
Add colredlogs to env yaml file
