+++ environment.yml
@@ -10,3 +10,4 @@ dependencies:
   - pytest=8.3.5
   - setuptools=80.1.0
   - click=8.1.8
+  - coloredlogs=15.0.1
