### 2025-05-23 ba48eac
Add extra help messaging in seabreeze CLI, update docs
