+++ documentation/docs/usage.md
@@ -1,28 +1,75 @@
 # Preparing input
 
-Assuming you have [completed installation](installation.md), you can now prepare input files for analysis. At its core, _seabreeze_ is designed for bacterial resequencing experiments, and therefore predicts mutations between reference-query pairs, where the reference sequence is referred to as an _ancestor_ and the query sequence being compared to is called the _assembly_. This nomenclature is use throughout the documentation and the software.
+Assuming you have [completed installation](installation.md), you can now prepare input files for analysis. At its core, _seabreeze_ is designed for bacterial resequencing experiments, and therefore predicts mutations between reference-query pairs, where the reference sequence is referred to as the _ancestor_ and the query sequence being compared to is called the _assembly_. This nomenclature is use throughout the documentation and the software.
 _seabreeze_ requires FASTA files for all of the bacterial genomes being analysed. These FASTA files should only contain a single contig for the full-length genome. It should not contain additional entries for plasmids. _seabreeze_ works best with genome assemblies generated from long-read sequencing as these can capture structural variants that occur between repeat sequences. It also does not work well for comparing genomes with high sequence/ phylogenetic divergence (for example between different bacterial species).
 
-All of the input/output of the analysis is stored in the `data/` directory in the _seabreeze_ folder. The pairwise comparisons to be performed are specified in a csv file called `data.csv` in the `data/` directory. This file has only two columns `ancestor` and `assembly`, in a standard csv format. Populate the csv file with the ancestor-assembly pairs you wish to analyse. For example:
+_seabreeze_ has two modes: 
+- `run`:  Analyzing one ancestor-assembly pair
+- `batch`: Analyzing multiple samples at once
+
+Both modes run the same workflows, but require input to be supplied differently. 
+## `run`: Analyzing one ancestor-assembly pair
+
+```
+Usage: seabreeze run [OPTIONS] [WORKFLOW] [SNAKEMAKE_ARGS]...
+
+Options:
+  -d, --dir PATH     Set working directory. Default is current working
+                     directory
+  --ancestor PATH    Path to FASTA file of the ancestor or reference sequence
+  --assembly PATH    Path to FASTA file of the assembly or query sequence
+  --threads INTEGER  Resources to be used.  [default: 4]
+  --ori TEXT         Sequence of the origin. Must use IUPAC bases, and be
+		             unique in both the ancestor and assembly.  Required for                          predict_replichore_balance or run_all workflows
+  --dif TEXT         Sequence of the terminus. Must use standard IUPAC bases,
+                     and be unique in both the ancestor and assembly. Required                        for predict_replichore_balance or run_all workflows
+  --masked           Mask insertion sequences
+  -h, --help         Show this message and exit.
+
+```
+
+- All input is directly supplied via the command line. No set up required.
+## `batch`: Analyzing multiple samples at once
+
+```
+Usage: seabreeze batch [OPTIONS] [WORKFLOW] [SNAKEMAKE_ARGS]...
+
+  Determine snakemake command to run based on supplied args
+
+Options:
+  -d, --dir PATH     Set working directory. Default is current working
+                     directory
+  -b, --data PATH    Path to data.csv file. Default is data.csv in current
+                     working dir
+  --oridif PATH      Path to csv file with sequences for the origin and
+                     terminus. Default is ori_dif_coords.csv in the current dir.
+                     Required for predict_replichore_balance or run_all workflows
+  --threads INTEGER  Resources to be used.  [default: 4]
+  --masked           Mask insertion sequences
+  -h, --help         Show this message and exit.
[…]
