+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "seabreeze"
-version = "1.3.0"
+version = "1.4.0"
 
 [tool.setuptools.packages.find]
 where = ["src"]

[…]
