+++ documentation/mkdocs.yml
@@ -5,3 +5,4 @@ nav:
   - Usage: usage.md
   - Output: output.md
   - Tutorial: tutorial.md
+  - Workflows: workflows.md
