### 2025-05-24 5b250fb
Improve CLI help messages, pin ISEscan version number, add pytest env, update contributing.md
