### 2025-05-24 barricklab/seabreeze#4 @ira-zibbu
Hi @dr-joe-wirth ,

I have added instructions for testing to `contributing.md`
