### 2025-05-24 barricklab/seabreeze#5 @ira-zibbu
Hi,

- Given that I have created a CLI wrapper around the Snakemake workflow, I have now added additional tests for the whole workflow (as opposed to only unit tests for individual scripts or packages). These test scripts are in `test/`, and use the standard pytest framework. Instructions for testing have been added to `contributing.md`

- I have also added these tests to the GitHub Actions, in addition to testing the `pip` build.

```
      - name: Build project from pyproject.toml
        run: |
          pip install .
        shell: bash -l {0}

      - name: Run specific tests with pytest
        run: |
          pytest test/seabreeze_*.py
        shell: bash -l {0}
```
