### 2025-05-28 321a1de
Change testing environment.yml to be minimally specified
