+++ test/unit_tests/environment.yml
@@ -3,153 +3,12 @@ channels:
   - conda-forge
   - bioconda
 dependencies:
-  - _libgcc_mutex=0.1=conda_forge
-  - _openmp_mutex=4.5=2_gnu
-  - amply=0.1.6=pyhd8ed1ab_0
-  - arpack=3.9.1=nompi_h77f6705_101
-  - biopython=1.84=py39hd3abc70_0
-  - blast=2.16.0=hc155240_3
-  - bzip2=1.0.8=h4bc722e_7
-  - c-ares=1.34.3=hb9d3cd8_1
-  - ca-certificates=2025.4.26=hbd8a1cb_0
-  - certifi=2025.1.31=pyhd8ed1ab_0
-  - click=8.1.8=pyh707e725_0
-  - coin-or-cbc=2.10.12=h8b142ea_1
-  - coin-or-cgl=0.60.9=h1d3f3f2_0
-  - coin-or-clp=1.17.10=h07f2a63_0
-  - coin-or-osi=0.108.11=h6514dde_1
-  - coin-or-utils=2.11.12=h99da652_1
-  - coincbc=2.10.12=1_metapackage
-  - colorama=0.4.6=pyhd8ed1ab_1
-  - curl=8.10.1=hbbe4b11_0
-  - cycler=0.12.1=pyhd8ed1ab_1
-  - docutils=0.21.2=pyhd8ed1ab_1
-  - entrez-direct=22.4=he881be0_0
-  - exceptiongroup=1.2.2=pyhd8ed1ab_1
-  - expat=2.6.4=h5888daf_0
-  - fastcluster=1.2.6=py39hec4eb1f_4
-  - fraggenescan=1.30=h470a237_2
-  - freetype=2.12.1=h267a509_2
-  - gdbm=1.18=h0a1914f_2
-  - gettext=0.22.5=he02047a_3
-  - gettext-tools=0.22.5=he02047a_3
-  - glpk=5.0=h445213a_0
-  - gmp=6.3.0=hac33072_2
-  - hmmer=3.4=hdbdd923_2
-  - iniconfig=2.0.0=pyhd8ed1ab_1
-  - isescan=1.7.2.3=h031d066_2
-  - keyutils=1.6.1=h166bdaf_0
-  - kiwisolver=1.4.7=py39h74842e3_0
-  - krb5=1.21.3=h659f571_0
-  - lcms2=2.16=hb7c19ff_0
-  - ld_impl_linux-64=2.43=h712a8e2_2
-  - lerc=4.0.0=h27087fc_0
-  - libasprintf=0.22.5=he8f35ee_3
-  - libasprintf-devel=0.22.5=he8f35ee_3
-  - libblas=3.9.0=25_linux64_openblas
-  - libcblas=3.9.0=25_linux64_openblas
-  - libcurl=8.10.1=hbbe4b11_0
-  - libdeflate=1.21=h4bc722e_0
-  - libedit=3.1.20191231=he28a2e2_2
-  - libev=4.33=hd590300_2
-  - libexpat=2.6.4=h5888daf_0
-  - libffi=3.4.2=h7f98852_5
-  - libgcc=14.2.0=h77fa898_1
-  - libgcc-ng=14.2.0=h69a702a_1
-  - libgettextpo=0.22.5=he02047a_3
-  - libgettextpo-devel=0.22.5=he02047a_3
-  - libgfortran=14.2.0=h69a702a_1
-  - libgfortran-ng=14.2.0=h69a702a_1
-  - libgfortran5=14.2.0=hd5240d6_1
-  - libgomp=14.2.0=h77fa898_1
[…]
