### 2025-05-28 00b1b25
Update docs with install instructions
