### 2025-05-28 barricklab/seabreeze#4 @ira-zibbu
Hi @dr-joe-wirth,

Thanks for bringing this issue to my attention. I have updated the `environment.yml` to be minimally specified.
