### 2025-05-31 barricklab/seabreeze#4 @dr-joe-wirth
When I try creating the environment I get the following error:

```text
PackagesNotFoundError: The following packages are not available from current channels:

  - isescan=1.7.3
```

After some digging, I believe that `isescan` is not compatible with macos arm64 (M1/M2/M3 chips).
