### 2025-06-06 barricklab/seabreeze#7 @JeanMainguy
Hi,

Great to hear you submitted a PR, hopefully the authors will respond soon ! In any case, this doesn’t need to block the review process for now. Just make sure to update the repository if they accept the PR, so the codebase remains clean and maintainable.

Also, I noticed that the `__pycache__` directory from the `plotsr` module was committed to the repository. These files are automatically generated by Python and don’t need to be tracked in version control.

I ran into this because they were modified locally when I used the tool, and doing a simple `git pull` caused conflicts.
