### 2025-06-13 barricklab/seabreeze#7 @ira-zibbu
Hi @JeanMainguy ,

Oops, yes, those cache files shouldn't be tracked by git. I will remove them from the git history. Thanks for letting me know.
