### 2025-06-13 barricklab/seabreeze#5 @ira-zibbu
@JeanMainguy ,

Currently the CI is set up to run the top-level CLI wrapper, which includes running all of those individual scripts when Snakemake calls them. Would you still like for me to add the unit tests to the CI?

I will update all of the environment YAML files to remove the defaults channel. Thanks for bringing this to my notice.
