### 2025-06-13 0278299
Remove defaults as a channel from the yml files
