+++ src/seabreeze/workflow/envs/syri.yml
@@ -1,6 +1,5 @@
 channels:
   - conda-forge
   - bioconda
-  - defaults
 dependencies:
   - syri=1.6.3
