+++ src/seabreeze/workflow/envs/prokka.yml
@@ -1,7 +1,6 @@
 channels:
   - conda-forge
   - bioconda
-  - defaults
 dependencies:
   - prokka=1.14.6
   - breseq=0.38.1
