+++ src/seabreeze/workflow/envs/mummer4.yml
@@ -1,6 +1,5 @@
 channels:
   - conda-forge
   - bioconda
-  - defaults
 dependencies:
   - mummer=3.23
