+++ src/seabreeze/workflow/envs/isescan.yml
@@ -1,6 +1,5 @@
 channels:
   - conda-forge
   - bioconda
-  - defaults
 dependencies:
   - isescan=1.7.3
