+++ src/seabreeze/workflow/envs/biopython.yml
@@ -2,7 +2,6 @@ name: biopython
 channels:
   - conda-forge
   - bioconda
-  - defaults
 dependencies:
   - biopython
   - pandas=2.0.3
