+++ src/seabreeze/workflow/envs/pandas.yml
@@ -2,7 +2,6 @@ name: pandas
 channels:
   - conda-forge
   - bioconda
-  - defaults
 dependencies:
   - pandas=2.0.3
   - numpy=1.25.0
