+++ src/seabreeze/workflow/envs/seabreeze.yml
@@ -2,6 +2,5 @@ name: seabreeze
 channels:
   - conda-forge
   - bioconda
-  - defaults
 dependencies:
   - snakemake=7.32.4
