+++ src/seabreeze/workflow/envs/breseq.yml
@@ -1,7 +1,6 @@
 channels:
   - conda-forge
   - bioconda
-  - defaults
 dependencies:
   - breseq=0.38.1
   - pandas=2.0.3
