### 2025-06-16 barricklab/seabreeze#5 @JeanMainguy
Hi,

Great, thanks for removing the defaults channel!

Regarding the tests: yes, I think it’s generally good practice to include unit tests in the CI alongside the full workflow tests. Since you already have them written, it makes sense to include them. Unit tests provide fine-grained checks that can help catch issues early.
