### 2025-06-20 177d035
Add unit tests to CI workflow
