### 2025-06-20 barricklab/seabreeze#5 @ira-zibbu
Hi,

I have added the unit tests to the CI workflow. See below an excerpt from `ci.yaml` that specifies the workflow.

```

      - name: Run unit tests with pytest
        run: |
          pytest test/unit_tests/*.py
        shell: bash -l {0}
```
