### 2025-06-23 barricklab/seabreeze#4 @dr-joe-wirth
## TLDR
I ran the unit tests on a AWS linux server. your environment file is incomplete; `matplotlib` needs to be included as a dependency. see below for th

## failed tests

I ran the following commands

```bash
conda env create --name seabreeze-test --file ./seabreeze/test/unit_tests/environment.yml
conda activate seabreeze-test
pytest ./seabreeze/test/unit_tests/*py
```

this failed because it couldn't find files. i assumed that your unit tests had hardcoded a requirement that cwd is `./seabreeze`, so i did the following:

```bash
cd ./seabreeze
pytest ./test/unit_tests/*py
```

now this ran, but there were several failures:

```text
===================================== test session starts =====================================
rootdir: /local/home/joewirth/seabreeze-review/seabreeze
configfile: pytest.ini
collected 13 items

test/unit_tests/fast_stats_test.py .                                                    [  7%]
test/unit_tests/find_IS_elements_test.py F                                              [ 15%]
test/unit_tests/find_reindex_bases_test.py ..                                           [ 30%]
test/unit_tests/mask_IS_elements_test.py .                                              [ 38%]
test/unit_tests/nucmer_alignment_test.py .                                              [ 46%]
test/unit_tests/plotsr_test.py .F                                                       [ 61%]
test/unit_tests/reindex_assembl_test.py .                                               [ 69%]
test/unit_tests/rename_contigs_test.py .                                                [ 76%]
test/unit_tests/replichore_arms_analyse_test.py ..                                      [ 92%]
test/unit_tests/syri_call_variants_test.py .                                            [100%]
```

here are the relevant snippets of the failure messages

```text
========================================== FAILURES ===========================================
_________________________________________ test_isecan _________________________________________

... a lot of text ...

OSError: libssw.so: cannot open shared object file: No such file or directory
______________________________________ test_plotsr_plots ______________________________________

... a lot of text ...

ModuleNotFoundError: No module named 'matplotlib'
=================================== short test summary info ===================================
FAILED test/unit_tests/find_IS_elements_test.py::test_isecan - subprocess.CalledProcessError: Command '['isescan.py', '--seqfile', 'test/data/Anc-_0gen_R...
FAILED test/unit_tests/plotsr_test.py::test_plotsr_plots - AssertionError: assert False
================================ 2 failed, 11 passed in 3.11s =================================
```

## fixing the problem
I deleted the `conda` environment, and changed the environment file to the one below:

```yaml
name: seabreeze_test
channels:
  - conda-forge
  - bioconda
dependencies:
  - biopython=1.84
  - blast=2.16.0
  - isescan=1.7.3
  - matplotlib         # this is the only line i added
  - matplotlib-base=3.3.4
  - mummer=3.23
  - numpy=1.26.4
  - pandas=2.0.3
  - pytest=8.3.3
  - syri=1.6.3
```

then i rebuilt the `conda` environment and ran the tests. now all tests passing.
