### 2025-06-23 barricklab/seabreeze#9 @dr-joe-wirth
tutorial failed
I am using linux on an AWS instance. I ran the following commands:

```bash
git clone https://github.com/barricklab/seabreeze.git
cd ./seabreeze
conda env create -f ./environment.yml
conda activate seabreeze
pip install ./
cd ..
cp -r ./seabreeze/example ./example_data
seabreeze batch --dir ./example_data --data example_data/data.csv analyse_genome_sizes
```

and I got the following error message:

```text
2025-06-23 14:45:50 dev-dsk-joewirth-2a-52dcfcae.us-west-2.amazon.com seabreeze[61993] DEBUG ==================================
2025-06-23 14:45:50 dev-dsk-joewirth-2a-52dcfcae.us-west-2.amazon.com seabreeze[61993] DEBUG seabrezee version: 1.3.0
2025-06-23 14:45:50 dev-dsk-joewirth-2a-52dcfcae.us-west-2.amazon.com seabreeze[61993] DEBUG ==================================
2025-06-23 14:45:50 dev-dsk-joewirth-2a-52dcfcae.us-west-2.amazon.com seabreeze[61993] DEBUG Switch working directory to /local/home/joewirth/seabreeze-review/example_data
2025-06-23 14:45:50 dev-dsk-joewirth-2a-52dcfcae.us-west-2.amazon.com seabreeze[61993] DEBUG seabrezee version: 1.3.0
seabreeze package path is /home/joewirth/miniforge3/envs/seabreeze/lib/python3.11/site-packages/seabreeze
2025-06-23 14:45:50 dev-dsk-joewirth-2a-52dcfcae.us-west-2.amazon.com seabreeze[61993] DEBUG Executing: snakemake --snakefile /home/joewirth/miniforge3/envs/seabreeze/lib/python3.11/site-packages/seabreeze/Snakefile --use-conda --cores 4 analyse_genome_sizes --config data=/local/home/joewirth/seabreeze-review/example_data/data.csv oridif=/local/home/joewirth/seabreeze-review/ori_dif_coords.csv threads=4
Building DAG of jobs...
Creating conda environment /home/joewirth/miniforge3/envs/seabreeze/lib/python3.11/site-packages/seabreeze/workflow/envs/biopython.yml...
Downloading and installing remote packages.
CreateCondaEnvironmentException:
Could not create conda environment from /home/joewirth/miniforge3/envs/seabreeze/lib/python3.11/site-packages/seabreeze/workflow/envs/biopython.yml:
Command:
mamba env create --quiet --file "/local/home/joewirth/seabreeze-review/example_data/.snakemake/conda/9a3619c3c48b01f0ed252353eb3d8258_.yaml" --prefix "/local/home/joewirth/seabreeze-review/example_data/.snakemake/conda/9a3619c3c48b01f0ed252353eb3d8258_"
Output:
error    libmamba Non-conda folder exists at prefix
critical libmamba Aborting.

2025-06-23 14:45:51 dev-dsk-joewirth-2a-52dcfcae.us-west-2.amazon.com seabreeze[61993] CRITICAL Command 'snakemake --snakefile /home/joewirth/miniforge3/envs/seabreeze/lib/python3.11/site-packages/seabreeze/Snakefile --use-conda --cores 4 analyse_genome_sizes --config data=/local/home/joewirth/seabreeze-review/example_data/data.csv oridif=/local/home/joewirth/seabreeze-review/ori_dif_coords.csv threads=4  ' returned non-zero exit status 1.
```

this is related to https://github.com/openjournals/joss-reviews/issues/8065
