### 2025-06-24 barricklab/seabreeze#9 @ira-zibbu
Hi @dr-joe-wirth,

I also recently encountered this issue, and is a [known issue with mamba](https://github.com/mamba-org/mamba/issues/2736). I have documented in on the [seabreeze usage page](https://barricklab.github.io/seabreeze/usage/#additional-snakemake-argument).

Can you try appending the flag `--conda-frontend conda` to see if that resolves it?
