### 2025-06-24 f63017e
Update dependencies in the test env
