### 2025-06-24 barricklab/seabreeze#4 @ira-zibbu
Hi @dr-joe-wirth,

Sorry to hear about the trouble you have been having.

- Regarding ISEScan not running on arm64 MacOS: I have noted this issue on the documentation, as well the `contributing.md` to give users a heads up.

- Regarding the missing matplotlib dependency: I currently had only specified `matplotlib-base=3.3.4` as a dependency in the test environment. The unit tests ran fine on my system, as well as on GitHub Actions. I have updated the environment recipe to instead include `matplotlib` of `matplotlib-base=3.3.4`. This should resolve that missing dependency issue you encountered.

- Regarding failed tests: The step you ran was the workflow tests (as opposed to the unit tests). As specified in `contributing.md`, these tests need to be run from the seabreeze execution conda environment and not the seabreeze_test directory built for the unit tests. The seabreeze execution environment contains additional important dependencies. I have updated `contributing.md` to clarify this better.  Try:

```
conda activate seabreeze #assuming this is what your seabreeze execution env is named
pytest test/*.py
```

I have also updated `contributing.md` to indicate that the tests should be run from the seabreeze root directory in order for the test input file locations to be correctly found.

Let me know if you face any other issues, have questions or suggestions.
