### 2025-07-02 #8065 @JeanMainguy
Hi,

I've tested the tool again following the recent updates, and everything worked smoothly. The new additions significantly enhance the user experience. Well done @ira-zibbu !

I've completed the review checklist.
