### 2025-07-14 barricklab/seabreeze#4 @dr-joe-wirth
i installed seabreeze using:

```bash
conda create --name seabreeze-from-conda
conda activate seabreeze-from-conda
conda install seabreeze-genomics -c conda-forge -c bioconda
```

then i tried running the tests:

```bash
conda activate seabreeze-from-conda
cd ./seabreeze
pytest ./test/*py
```

and everything passed.
