### 2025-07-14 barricklab/seabreeze#4 @dr-joe-wirth
one note:

The gitignore file currently attempts to specify every possible path to `__pycache__`, but it does not catch all of them. Instead just add `__pycache__`

ie replace this:

```text
src/seabreeze/__pycache__/
src/seabreeze/plotsr/__pycache__
src/__pycache__/
src/seabreeze/plotsr/plotsr/scripts/__pycache__
```

with this:

```text
__pycache__
```

Running the tests creates a bunch of `__pycache__` directories and it would be nice if `git` ignored them.
