### 2025-07-14 barricklab/seabreeze#4 @ira-zibbu
Hi @dr-joe-wirth I have updated the `.gitignore` file.
