### 2025-07-16 #8065 @ira-zibbu
- [x] Double check Paper authors and affiliations (including ORCIDs)
- [x] Make a tagged release of your software on Github, and list the version tag of the archived version here.

- [seabreeze v1.5.0](https://github.com/barricklab/seabreeze/releases/tag/v1.5.0) corresponds to the archived version.

- [x] Check the archival deposit (e.g., in Zenodo) has the correct title (should match the paper title).
- [x] Check the archival deposit (e.g., in Zenodo) has the correct authors and affiliations (should match the paper authors including order).
- [x] Check the archival deposit (e.g., in Zenodo) has the correct version (should match the above Github tag you created).
- [x] Check the archival deposit (e.g., in Zenodo) has the correct ORCIDs (should match the paper authors).

https://doi.org/10.5281/zenodo.15994088
