### 2025-07-17 #8065 @abhishektiwari
@ira-zibbu Looks like you have reference `carver_act_2005` appears twice in the `paper.bib`. Can you fix that so we can generate PDF and XML Metadata again?
