### 2025-07-17 ad25485
Remove redudant ref in paper.bib
