+++ JOSS/paper.bib
@@ -308,23 +308,6 @@ Publisher: Cold Spring Harbor Lab},
 	langid = {english},
 }
 
-@article{carver_act_2005,
-	title = {{ACT}: the Artemis comparison tool},
-	volume = {21},
-	issn = {1367-4803, 1460-2059},
-	url = {https://academic.oup.com/bioinformatics/article-lookup/doi/10.1093/bioinformatics/bti553},
-	doi = {10.1093/bioinformatics/bti553},
-	shorttitle = {{ACT}},
-	pages = {3422--3423},
-	number = {16},
-	journaltitle = {Bioinformatics},
-	shortjournal = {Bioinformatics},
-	author = {Carver, T. J. and Rutherford, K. M. and Berriman, M. and Rajandream, M.-A. and Barrell, B. G. and Parkhill, J.},
-	urldate = {2025-01-31},
-	date = {2005-08-15},
-	langid = {english},
-}
-
 
 @article{achaz_associations_2003,
 	title = {Associations Between Inverted Repeats and the Structural Evolution of Bacterial Genomes},
