### 2025-07-24 0d81cd0
Add author affiliation locations in paper.md
