### 2025-07-24 #8065 @ira-zibbu
Looks great! Thank you everyone.

Just one quick note: seems like the italics in the title is not being rendered correctly. Is there a way to resolve this? It is also okay with the authors to remove the markup altogether.
