### 2025-07-25 38defce
Remove italics markup in title in paper.md, and add citation.cff
