### 2025-07-25 #8065 @ira-zibbu
I have updated the title in `paper.md` on my repository.
